BlackList refactored.
