--- src/main/java/ir/niopdc/simulator/DCSimulatorApplication.java
@@ -1,17 +1,0 @@
-package ir.niopdc.simulator;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.EnableAspectJAutoProxy;
-import org.springframework.scheduling.annotation.EnableScheduling;
-
-@SpringBootApplication
-@EnableAspectJAutoProxy
-@EnableScheduling
-public class DCSimulatorApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(DCSimulatorApplication.class, args);
-	}
-
-}
