--- src/main/java/ir/niopdc/simulator/constant/ErrorMessages.java
@@ -1,7 +1,0 @@
-package ir.niopdc.simulator.constant;
-
-public final class ErrorMessages {
-    private ErrorMessages() {}
-
-    public static final String POLICY_NOT_FOUND = "Policy not found for %s";
-}
