+++ src/main/resources/application.properties
@@ -1,3 +1,4 @@
 spring.profiles.active=dev
 spring.application.name=DC-Simulator
-grpc.server.port=9090
+
+app.config.cron.black-list = 0 */1 * * * ?
